requirements.txt
fastapi==0.101.1
openai==1.10.0
python-dotenv
uvicorn==0.23.2